General cleanups and database generalization

+++ src/TimeFormatter.php
@@ -6,10 +6,10 @@ final class TimeFormatter
 {
     public static function format(int $time): string
     {
-        $tstring = date('H:i:s', $time / 1000);
+        $tstring = date('H:i:s', (int) ($time / 1000));
         $tstring = ltrim($tstring, '0: ');
         $tstring .= '.'.substr($time, -3, 2);
-        $day = date('j', $time / 1000) - 1;
+        $day = date('j', (int) ($time / 1000)) - 1;
         if ($day) {
             $tstring = $day.'d '.$tstring;
         }
